fix: dockerUtils file added
